refactor(tests): extract testing env vars to constants
